Refactor Server class separate file
Move Server class to factorio_server.js and rename it to FactorioServer.
Refactor out the call directly to the server side and the direct usage
of the Airtable base.
