+++ airtable.js
@@ -2,9 +2,11 @@
 const { lua_array } = require('./helpers.js');
 
 
+let base;
+
 //Get the id for the record of the player by the given name in the Player Data table
 let player_cache = new Map();
-async function get_player_id(base, name) {
+async function get_player_id(name) {
     if (player_cache.has(name)) {
         return player_cache.get(name);
     }
@@ -25,7 +27,7 @@ async function get_player_id(base, name) {
 
 //Get the id for the record of the game by the given name in the Game Internals table
 let game_cache = new Map();
-async function get_game_id(base, name) {
+async function get_game_id(name) {
     if (game_cache.has(name)) {
         return game_cache.get(name);
     }
@@ -45,18 +47,18 @@ async function get_game_id(base, name) {
 };
 
 //Create match record in Scoring Data table.
-exports.started_game = async function(base, name, players) {
+exports.started_game = async function(name, players) {
     let fields = {};
     fields["Players Present"] = [];
     fields["Time Started"] = new Date().toISOString();
-    let game_id = await get_game_id(base, name);
+    let game_id = await get_game_id(name);
     if (game_id !== null) {
         fields["Game"] = [game_id];
     } else {
         console.log(`Warning: Got started_game for nonexistent game ${name}`);
     }
     for (let player of players) {
-        let player_id = await get_player_id(base, player);
+        let player_id = await get_player_id(player);
         if (player_id) {
             fields["Players Present"].push(player_id);
         }
@@ -68,7 +70,7 @@ exports.started_game = async function(base, name, players) {
 };
 
 //Update match record in Scoring Data table with pole positions.
-exports.stopped_game = async function(base, results, record_id) {
+exports.stopped_game = async function(results, record_id) {
     let fields = {};
 
     let place_to_field = [null, "Gold", "Silver", "Bronze"];
@@ -78,7 +80,7 @@ exports.stopped_game = async function(base, results, record_id) {
             continue;
         }
         let player_ids = await Promise.all(
-            lua_array(entry.players).map(player => get_player_id(base, player))
+            lua_array(entry.players).map(player => get_player_id(player))
         );
         fields[`${place_to_field[entry.place]} Player`] = player_ids.filter((id) => id !== null);
         fields[`${place_to_field[entry.place]} Data`] = entry.score;
@@ -89,8 +91,8 @@ exports.stopped_game = async function(base, results, record_id) {
     console.log(await base('Scoring Data').update(record_id, fields));
 };
 
-exports.add_player = async function(base, player_name) {
-    let player_id = await get_player_id(base, player_name);
+exports.add_player = async function(player_name) {
+    let player_id = await get_player_id(player_name);
     if (player_id !== null) { return; };
     const player_record = await base('Player Data').create({
         "Player Name": `${player_name}`,
@@ -114,7 +116,7 @@ airtable_events.on("newListener", function (event, listener) {
     }
 });
 
-async function check_roles(base) {
+async function check_roles() {
     let check = new Date().toISOString();
     await base('Player Data').select({
         fields: ["Roles", "Player Name"],
@@ -150,8 +152,11 @@ async function check_roles(base) {
 
 exports.airtable_events = airtable_events;
 exports.player_roles = players_roles;
-exports.init = async function init(base) {
+exports.set_base = function set_base(new_base) {
+    base = new_base;
+};
 
+exports.init = async function init() {
     last_checked = new Date().toISOString();
     let pages = await base('Player Data').select({
         fields: ["Roles", "Player Name"],
@@ -170,7 +175,7 @@ exports.init = async function init(base) {
         });
     airtable_events.emit('init', players_roles);
     setInterval(() => {
-        check_roles(base).catch((err) => {
+        check_roles().catch((err) => {
             console.error(err);
         });
     }, 10000);
